legacy-logging; interim; failing tests
