+++ lib/common.js
@@ -239,6 +239,8 @@ exports.makedie = function (instance, ctxt) {
   ctxt = _.extend(ctxt, instance.die ? instance.die.context : {})
 
   var die = function (err) {
+    console.log(err.stack)
+
     var die_trace = '\n' + (new Error('die trace').stack)
         .match(/^.*?\n.*\n(.*)/)[1]
