+++ lib/legacy.js
@@ -62,7 +62,9 @@ var start_time = Date.now()
 
 handlers.pretty = function () {
   var args = arrayify(arguments)
-  args[2] = args[2].toUpperCase()
+  if( _.isString(args[2]) ) {
+    args[2] = args[2].toUpperCase()
+  }
 
   if (args[0].short$) {
     args[0] = '' + (args[0].getTime() - start_time)
@@ -113,3 +115,111 @@ handlers.file = function file (filepath, opts) {
   var ws = Fs.createWriteStream(filepath, {flags: opts.flags || 'a'})
   return handlers.stream(ws, opts)
 }
+
+
+function parse_command_line (spec, logspec, flags) {
+  flags = flags || {}
+
+  var logmaps = logspec.map
+
+  if (flags.shortcut) {
+    if (spec === 'short' || spec.short === true) {
+      logspec.short = true
+      if (!logmaps.length) {
+        logmaps.push({ level: 'info+', handler: 'print' })
+      }
+    }
+
+    var shortentries = logging_shortcut(spec)
+
+    if (shortentries.length) {
+      shortentries.forEach(function (shortentry) {
+        logmaps.push(shortentry)
+      })
+      return
+    }
+  }
+
+  if (_.isArray(spec)) {
+    spec.forEach(function (specentry) {
+      parse_command_line(specentry, logspec)
+    })
+    return
+  }
+
+  // parse: level=,type=,plugin=,tag=,case=,handler=
+  // handler can be print,file:path
+
+  var keys = { level: 1, type: 1, plugin: 1, tag: 1, 'case': 1,
+    handler: 1, regex: 1, pin: 1, act: 1 }
+  var entry = {}
+  var parts = ('' + spec).split(',')
+  _.each(parts, function (part) {
+    var kvm = part.match(/^(.*?):(.*)$/)
+    var kv = kvm ? [kvm[1], kvm[2]] : ['']
+
+    if (kv[0].length) {
+      var key = kv[0]
+      if (key === 'handler') {
+        var handler = kv.slice(1).join(':')
+        var m
+        if (handler === 'print') {
+          entry[key] = handlers.print
+        }
+        else if ((m = /^file:(\/\/)?(.*)$/.exec(handler))) {
+          entry[key] = handlers.file(m[2])
+        }
+      }
+      else if (keys[key]) {
+        if (entry[key]) {
+          entry[key] += ',' + kv[1]
+        }
+        else {
+          entry[key] = kv[1]
+        }
+      }
+    }
+  })
+
+  if (_.keys(entry).length) {
+    // print by default
+    if (entry && !entry.handler) {
+      entry.handler = handlers.print
+    }
+
+    logmaps.push(entry)
+  }
+}
+
+exports.parse_command_line = parse_command_line
+
+
+function logging_shortcut (spec) {
+  if (spec && (spec.print === true ||
+    spec.all === true ||
+    spec === 'print' ||
+    spec === 'all')) {
+    return [{ level: 'all', handler: handlers.print }]
+  }
+  else if (spec &&
+    (spec.quiet ||
+    spec === 'quiet' ||
+    spec.silent ||
+    spec === 'silent')) {
+    return []
+  }
+  else if (spec === 'test') {
+    return [{ level: 'error+', handler: handlers.print }]
+  }
+  else if (spec === 'standard') {
+    return [{ level: 'info+', handler: handlers.print }]
+  }
+  else if (_.isString(spec)) {
+    var logspec = {map: []}
+    parse_command_line(spec, logspec, {shortcut: false})
+    return logspec.map
+  }
+  else return []
+}
+
+exports.logging_shortcut = logging_shortcut
