+++ lib/optioner.js
@@ -1,4 +1,4 @@
-/* Copyright (c) 2014-2015 Richard Rodger, MIT License */
+/* Copyright (c) 2014-2016 Richard Rodger and other contributors, MIT License */
 'use strict'
 
 var Fs = require('fs')
@@ -6,7 +6,7 @@ var _ = require('lodash')
 var Error = require('eraro')
 var Jsonic = require('jsonic')
 var Minimist = require('minimist')
-var Logging = require('./logging')
+var Legacy = require('./legacy')
 var Common = require('./common')
 
 var error = Error({ package: 'seneca', msgmap: ERRMSGMAP() })
@@ -51,7 +51,7 @@ module.exports = function (callmodule, defaults) {
   if (process.env.SENECA_LOG) {
     sourcemap.env.log = sourcemap.env.log || {}
     sourcemap.env.log.map = sourcemap.env.log.map || []
-    Logging.parse_command_line(process.env.SENECA_LOG,
+    Legacy.parse_command_line(process.env.SENECA_LOG,
       sourcemap.env.log,
       {shortcut: true})
   }
@@ -86,7 +86,7 @@ module.exports = function (callmodule, defaults) {
     if (argv.seneca.log) {
       sourcemap.argv.log = sourcemap.argv.log || {}
       sourcemap.argv.log.map = sourcemap.argv.log.map || []
-      Logging.parse_command_line(argv.seneca.log,
+      Legacy.parse_command_line(argv.seneca.log,
         sourcemap.argv.log,
         { shortcut: true })
     }
