+++ lib/plugins.js
@@ -6,7 +6,6 @@ var _ = require('lodash')
 var Eraro = require('eraro')
 var Parambulator = require('parambulator')
 var Common = require('./common')
-var Logging = require('./logging')
 
 // Shortcuts
 var arrayify = Function.prototype.apply.bind(Array.prototype.slice)
@@ -242,6 +241,16 @@ function make_delegate (instance, plugin) {
     fatal$: true
   })
 
+
+  delegate.log = instance.make_log(
+    delegate, 
+    function plugin_delegate_log_modifier (data) {
+      data.plugin_name = plugin.name
+      data.plugin_tag = plugin.tag
+    })
+
+
+  /*
   delegate.log = function (level) {
     var args = arrayify(arguments)
 
@@ -256,6 +265,10 @@ function make_delegate (instance, plugin) {
     instance.log.apply(instance, args)
   }
   Logging.makelogfuncs(delegate)
+   */
+
+  
+
 
   delegate.die = Common.makedie(delegate, { type: 'plugin', plugin: plugin.name })
 
@@ -346,6 +359,17 @@ function resolve_plugin_exports (seneca, pluginref, meta) {
   if (_.isObject(meta.extend)) {
     if (_.isFunction(meta.extend.action_modifier)) {
       seneca.private$.action_modifiers.push(meta.extend.action_modifier)
+    }
+
+    if (_.isFunction(meta.extend.logger)) {
+      if (!meta.extend.logger.replace && 
+          _.isFunction(seneca.private$.logger.add)) 
+      {
+        seneca.private$.logger.add(meta.extend.logger)
+      }
+      else {
+        seneca.private$.logger = meta.extend.logger
+      }
     }
   }
